Use JNI loading to locate CUDA install instead of OS-specific files
Is platform-agnostic and doesn't need to hardcode install locations - CUDA just needs to be on the PATH
